Switched file size from Integer to Long.

+++ seqware-common/src/main/java/net/sourceforge/seqware/common/metadata/MetadataDB.java
@@ -928,7 +928,7 @@ public class MetadataDB extends Metadata {
     return value == null ? "null" : "'" + value + "'";
   }
   
-  private String sqlQuote(Integer value) {
+  private String sqlQuote(Long value) {
      return value == null ? "null" : value.toString();
    }
